requirements.txt
google-genai
openai-whisper
ffmpeg-python